evaluate/CheXStruct/requirements.txt
imageio==2.37.0
lazy_loader==0.4
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parmap==1.7.0
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
tifffile==2024.8.30
tzdata==2025.2
